+++ examples/rust/src/bin/tx-blocktime.rs
@@ -1,5 +1,5 @@
 use {
-    chrono::{DateTime, NaiveDateTime, Utc},
+    chrono::{DateTime, Utc},
     clap::{Parser, ValueEnum},
     futures::{sink::SinkExt, stream::StreamExt},
     log::{error, info},
@@ -131,10 +131,8 @@ async fn main() -> anyhow::Result<()> {
                     Some(UpdateOneof::BlockMeta(block)) => {
                         let entry = messages.entry(block.slot).or_default();
                         entry.0 = block.block_time.map(|obj| {
-                            DateTime::from_naive_utc_and_offset(
-                                NaiveDateTime::from_timestamp_opt(obj.timestamp, 0).unwrap(),
-                                Utc,
-                            )
+                            DateTime::from_timestamp(obj.timestamp, 0)
+                                .expect("invalid or out-of-range datetime")
                         });
                         if let Some(timestamp) = entry.0 {
                             for sig in &entry.1 {