+++ yellowstone-grpc-geyser/src/config.rs
@@ -1,8 +1,8 @@
 use {
-    serde::{de, Deserialize, Deserializer},
-    solana_geyser_plugin_interface::geyser_plugin_interface::{
+    agave_geyser_plugin_interface::geyser_plugin_interface::{
         GeyserPluginError, Result as PluginResult,
     },
+    serde::{de, Deserialize, Deserializer},
     solana_sdk::pubkey::Pubkey,
     std::{collections::HashSet, fs::read_to_string, net::SocketAddr, path::Path},
     tokio::sync::Semaphore,